Added SimpleFTPClient class
This class inherits from the abstract BaseFTPClient class and provides a wrapper over the most commonly used ftp functions. If some function is not available in this class, it is possible to use the public method called ftp from the abstract class through which the missing ftp function can be activate.